[ feat ] 커뮤니티 카테고리 컴포넌트 화
